refactor: split 3 oversized source files to restore <=500 LOC gate
