+++ memory-core/src/embeddings/mod.rs
@@ -52,6 +52,7 @@ mod provider;
 #[cfg(feature = "local-embeddings")]
 mod real_model;
 mod semantic_service;
+mod semantic_text;
 mod similarity;
 pub mod simple;
 mod storage;
